tests/wpt/web-platform-tests/html/semantics/embedded-content/the-canvas-element/historical.html

<!doctype html>
<meta charset=utf-8>
<title>Historical canvas features</title>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<div id=log></div>
<script>
var canvas;
setup(function() {
  canvas = document.createElement("canvas");
});
["supportsContext", "probablySupportsContext"].forEach(function(m) {
  test(function() {
    assert_false(m in canvas);
  }, "Canvas support for " + m);
});
</script>
